Extract SessionForm into a component
